Add functionality to calculate intake based on imperial units
The program now asks the user in the beginning for their preffered measurement
system of choice (metric / imperial). Based on that, the prompts for height
and weight differ. If the user chooses metric, the program calculates everything
as before. If the user chooses imperial, the program asks the user for their
height in feet and inches, and their weight in pounds. The program then converts
the imperial units into metric units and continues as before.